Refactor: move stream reader and writer to pixels-core
